Chore: Reorganize ee hooks
