+++ packages/core/admin/ee/admin/hooks/__mocks__/useLicenseLimits.js
