--- packages/core/admin/ee/admin/hooks/index.js
@@ -1,4 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export { default as useAuthProviders } from './useAuthProviders';
-export { default as useLicenseLimitNotification } from './useLicenseLimitNotification';
-export { useLicenseLimits } from './useLicenseLimits';
