+++ packages/core/admin/ee/admin/hooks/tests/useLicenseLimitNotification.js
@@ -3,8 +3,8 @@ import React from 'react';
 import { renderHook } from '@testing-library/react';
 import { IntlProvider } from 'react-intl';
 
-import useLicenseLimitNotification from '..';
-import { useLicenseLimits } from '../../useLicenseLimits';
+import { useLicenseLimitNotification } from '../useLicenseLimitNotification';
+import { useLicenseLimits } from '../useLicenseLimits';
 
 const baseLicenseInfo = {
   enforcementUserCount: 5,
