--- packages/core/admin/ee/admin/hooks/useLicenseLimits/index.js
@@ -1,1 +1,0 @@
-export { useLicenseLimits } from './useLicenseLimits';
