+++ packages/core/admin/ee/admin/pages/HomePage/index.js
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 // eslint-disable-next-line import/no-cycle
 import { HomePageCE } from '../../../../admin/src/pages/HomePage';
-import { useLicenseLimitNotification } from '../../hooks';
+import { useLicenseLimitNotification } from '../../hooks/useLicenseLimitNotification';
 
 export function HomePageEE() {
   useLicenseLimitNotification();
