+++ packages/core/admin/ee/admin/pages/SettingsPage/pages/Users/ListPage/index.js
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 // eslint-disable-next-line import/no-cycle
 import { UserListPageCE } from '../../../../../../../admin/src/pages/SettingsPage/pages/Users/ListPage';
-import { useLicenseLimitNotification } from '../../../../../hooks';
+import { useLicenseLimitNotification } from '../../../../../hooks/useLicenseLimitNotification';
 
 function UserListPageEE() {
   useLicenseLimitNotification();
